Consolidate HTTP validation and common utilities into types module
• Move validation functions from validation.rs to types.rs
• Move parse_headers from common.rs to types.rs
• Remove validation.rs and common.rs files
• Update imports in cli.rs, server.rs, and mod.rs
