+++ src/http/mod.rs
@@ -6,11 +6,8 @@
 //! - Server implementation utilities
 
 pub mod types;
-pub mod validation;
 pub mod client;
 pub mod server;
-pub mod common;
 pub mod cli;
 
 pub use types::*;
-pub use validation::*;
