+++ src/http/server.rs
@@ -10,7 +10,6 @@ use axum::{
 use chrono::Utc;
 use crate::{FlashQ, FlashQError, Record};
 use super::types::*;
-use super::validation::*;
 use std::sync::Arc;
 use tokio::net::TcpListener;
